requirements.txt
django==3.2.9
django-cors-headers
paho-mqtt
Pillow
numpy
django-filter
djangorestframework
django-colorfield
drf-spectacular
djangorestframework-simplejwt
channels
channels-redis
mysqlclient
apscheduler
opencv-python